+++ alkemySpring/src/main/java/com/alfonso/alkemy/interfaces/IUsuarioService.java
@@ -8,8 +8,8 @@ import com.alfonso.alkemy.entity.Usuario;
 public interface IUsuarioService {
 
 	public List<Usuario> listUsuarios(String rol);
-	public Usuario getUsuario(Long id);
+	public Usuario getUsuario(Integer id);
 	public Usuario guardarUsuario(Usuario alumno);
 	public Usuario updateUsuario(Usuario alumno);
-	public void delete(Long id);
+	public void delete(Integer id);
 }
