+++ alkemySpring/src/main/java/com/alfonso/alkemy/repository/UsuariosRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.alfonso.alkemy.entity.Usuario;
 
-public interface UsuariosRepository extends JpaRepository<Usuario, Long>{
+public interface UsuariosRepository extends JpaRepository<Usuario, Integer>{
 
 	public List<Usuario> findByRol(String rol);
 }
